requirements.txt
requests
boto3==1.3.1
psycopg2-binary
